fix(browser): align remote stream modules with split pane
