+++ src/renderer/src/components/browser-pane/stream-remote/remote-browser-fake-screencast-stream.ts
