+++ src/renderer/src/components/browser-pane/stream-remote/remote-browser-first-frame-recovery.ts
