+++ src/renderer/src/components/browser-pane/stream-remote/remote-browser-frame-point.test.ts
