+++ src/renderer/src/components/browser-pane/stream-remote/remote-browser-legacy-viewport.test.ts
