+++ src/renderer/src/components/browser-pane/stream-remote/remote-browser-stream-events.ts
