+++ src/renderer/src/components/browser-pane/stream-remote/remote-browser-stream-first-frame.test.ts
