+++ src/renderer/src/components/browser-pane/stream-remote/use-remote-browser-stream-activation.test.ts
@@ -18,14 +18,12 @@ function renderActivation(isActive = true) {
   const closeStream = vi.fn()
   const open = vi.fn(() => closeStream)
   const lifecycle = { open }
-  const clearPendingRemoteInput = vi.fn()
   const clearPendingRemoteWheel = vi.fn()
   const hook = renderHook(
     ({ active }) =>
       useRemoteBrowserStreamActivation({
         activeRuntimeEnvironmentId: 'env-1',
         browserPageId: 'page-1',
-        clearPendingRemoteInput,
         clearPendingRemoteWheel,
         isActive: active,
         lifecycle,
@@ -34,7 +32,7 @@ function renderActivation(isActive = true) {
       }),
     { initialProps: { active: isActive } }
   )
-  return { ...hook, clearPendingRemoteInput, clearPendingRemoteWheel, closeStream, open }
+  return { ...hook, clearPendingRemoteWheel, closeStream, open }
 }
 
 describe('useRemoteBrowserStreamActivation', () => {
@@ -62,7 +60,6 @@ describe('useRemoteBrowserStreamActivation', () => {
 
     await act(async () => vi.advanceTimersByTime(1))
     expect(harness.closeStream).toHaveBeenCalledTimes(1)
-    expect(harness.clearPendingRemoteInput).toHaveBeenCalledTimes(1)
     expect(harness.clearPendingRemoteWheel).toHaveBeenCalledTimes(1)
 
     await act(async () => {
@@ -98,7 +95,6 @@ describe('useRemoteBrowserStreamActivation', () => {
 
     await act(async () => harness.rerender({ active: false }))
     expect(harness.closeStream).toHaveBeenCalledTimes(1)
-    expect(harness.clearPendingRemoteInput).toHaveBeenCalledTimes(1)
     expect(harness.clearPendingRemoteWheel).toHaveBeenCalledTimes(1)
 
     await act(async () => harness.rerender({ active: true }))
